Cherry pick configuration file of servers
