+++ smart-contracts/application-typescript/resources/certificate-body/config.json
@@ -1,0 +1,14 @@
+{
+  "version": "0.0.1",
+  "interface": "certificate-body",
+  "configuration": {
+    "chaincodeName": "basic",
+    "channelName": "mychannel",
+    "caName": "ca.org2.example.com",
+    "department": "org2.department1",
+    "filePath": "../../../../test-network/organizations/peerOrganizations/org2.example.com/connection-org2.json",
+    "mspOrg": "Org2MSP",
+    "userId": "appUser",
+    "walletPath": "wallet/walletCert"
+  }
+}
